WIP — Separar los modelos de la BDD para facilitar razonamiento ❤️😅 (#45)
"separation of concerns"/"separación de preocupaciones" depende mucho de
cuales sean tus preocupaciones 😅

Posterior a la apertura del proyecto y dada las conversas q tuvimos con
@Benjvvp, @Fabianko y @panchocorderos, le di una vuelta a los modelos de
la BDD para que nos enfoquemos más en su relación con la lógica de
negocios.

Esto es la parte 1, para limpiar y facilitar el análisis, antes de que
nos enfoquemos en arreglar el bloqueante de #41.

---------------------

PS: Esto ademas arregla el #43 que elimina la relación duplicada, y de
paso arregla la nombres de tablas q no estaban en `snake_case` 🐍.
https://github.com/JSConfCL/gql_api/pull/45/files#diff-7c562332aeb513611044fefea24792528452e9e029a3ecb1ca83fe1f94751e20R1-R17
