+++ .prettierignore
@@ -1,1 +1,2 @@
 *.generated.ts
+*.sql
