+++ src/datasources/db/schema.ts
@@ -1,3 +1,13 @@
-export * from "~/datasources/db/tables";
-export * from "~/datasources/db/relations";
-export * from "~/datasources/db/CRUD";
+// TABLES/RELATIONS/CRUD
+export * from "~/datasources/db/users";
+export * from "~/datasources/db/tickets";
+export * from "~/datasources/db/userTickets";
+export * from "~/datasources/db/communities";
+export * from "~/datasources/db/tags";
+export * from "~/datasources/db/events";
+export * from "~/datasources/db/allowedCurrencies";
+export * from "~/datasources/db/tagsCommunities";
+export * from "~/datasources/db/usersCommunities";
+export * from "~/datasources/db/eventsCommunities";
+export * from "~/datasources/db/eventsTags";
+export * from "~/datasources/db/eventsUsers";
