+++ src/datasources/queries/users.ts
@@ -1,8 +1,7 @@
 import { eq, sql } from "drizzle-orm";
 import { z } from "zod";
 import { ORM_TYPE } from "~/datasources/db";
-import { selectUsersSchema } from "~/datasources/db/schema";
-import { usersSchema } from "~/datasources/db/tables";
+import { selectUsersSchema, usersSchema } from "~/datasources/db/schema";
 import { getUsername } from "~/datasources/queries/utils/createUsername";
 
 export const ProfileInfoSchema = z.object({
