+++ src/schema/events.ts
@@ -1,4 +1,4 @@
-import { SQL, eq, gte, like, lte } from "drizzle-orm";
+import { SQL, eq, gte, like, lte, inArray } from "drizzle-orm";
 import { v4 } from "uuid";
 import { builder } from "~/builder";
 import {
@@ -12,8 +12,19 @@ import {
   selectUsersSchema,
   userTicketsSchema,
 } from "~/datasources/db/schema";
-import { CommunityRef, EventRef, TagRef, UserRef, UserTicketRef } from "~/schema/shared/refs";
-import { TicketApprovalStatus, TicketPaymentStatus, TicketRedemptionStatus, TicketStatus } from "./userTickets";
+import {
+  CommunityRef,
+  EventRef,
+  TagRef,
+  UserRef,
+  UserTicketRef,
+} from "~/schema/shared/refs";
+import {
+  TicketApprovalStatus,
+  TicketPaymentStatus,
+  TicketRedemptionStatus,
+  TicketStatus,
+} from "./userTickets";
 
 export const EventStatus = builder.enumType("EventStatus", {
   values: ["active", "inactive"] as const,
@@ -116,13 +127,8 @@ builder.objectType(EventRef, {
         input: t.arg({ type: EventsTicketsSearchInput, required: false }),
       },
       resolve: async (root, { input }, ctx) => {
-        const {
-          id,
-          status,
-          paymentStatus,
-          approvalStatus,
-          redemptionStatus
-        } = input ?? {};
+        const { id, status, paymentStatus, approvalStatus, redemptionStatus } =
+          input ?? {};
         const wheres: SQL[] = [];
         if (id) {
           wheres.push(eq(userTicketsSchema.id, id));
@@ -139,12 +145,39 @@ builder.objectType(EventRef, {
         if (redemptionStatus) {
           wheres.push(eq(userTicketsSchema.redemptionStatus, redemptionStatus));
         }
+
+        // TODO: (Felipe) — Esta es otra manera de hacerlo, aun no se cual es
+        // mejor. La diferencia es que con esta query, no es facil obtener un
+        // tipado fuerte en comparación a las relational queries.
+        // https://orm.drizzle.team/docs/rqb
+        //
+        // Es probable que podamos hacerlo si usamos el builder de drizzle-orm
+        // para generar el schema de la query, pero aun no lo he intentado.
+        //
+        // const withJoin = await ctx.DB.select()
+        //   .from(eventsSchema)
+        //   .innerJoin(ticketsSchema, eq(eventsSchema.id, ticketsSchema.eventId))
+        //   .innerJoin(
+        //     userTicketsSchema,
+        //     eq(ticketsSchema.id, userTicketsSchema.ticketTemplateId),
+        //   )
+        //   .where(eq(eventsSchema.id, root.id))
+        //   .run();
+
+        const ticketsTemplates = await ctx.DB.query.ticketsSchema.findMany({
+          where: (c, { eq }) => eq(c.eventId, root.id),
+        });
+        const ticketTemplateIds = ticketsTemplates.map((t) => t.id);
+        wheres.push(
+          inArray(userTicketsSchema.ticketTemplateId, ticketTemplateIds),
+        );
         const tickets = await ctx.DB.query.userTicketsSchema.findMany({
           where: (c, { and }) => and(...wheres),
           orderBy(fields, operators) {
             return operators.desc(fields.createdAt);
           },
         });
+
         return tickets.map((t) => selectUserTicketsSchema.parse(t));
       },
     }),
